added subscriber

+++ app/Http/Controllers/Author/PostController.php
@@ -97,6 +97,11 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('Sorry, you are not allowed to access this post !','Error');
+            return redirect()->back();
+        }
         return view('author.post.show',compact('post'));
     }
 
@@ -108,6 +113,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('Sorry, you are not allowed to access this post !','Error');
+            return redirect()->back();
+        }
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit',compact('post','categories','tags'));
@@ -122,6 +132,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('Sorry, you are not allowed to access this post !','Error');
+            return redirect()->back();
+        }
         $this->validate($request,[
            'title' => 'required',
            'image' => 'image',
@@ -180,6 +195,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id != Auth::id())
+        {
+            Toastr::error('Sorry, you are not allowed to access this post !','Error');
+            return redirect()->back();
+        }
         if(Storage::disk('public')->exists('post/'.$post->image))
         {
             Storage::disk('public')->delete('post/'.$post->image);
